:page_facing_up: Update the project's license from MIT to AGPL (Affero General Public License)
The decision to switch licenses aligns with our goals of promoting openness and ensuring that modifications and enhancements to the codebase remain accessible to the community. The AGPL provides additional protections, especially in the context of online services, and reinforces our commitment to transparency and collaboration. The complete text of the AGPL license has been included in the project for reference.